feat(k8s): Adds a helm chart (#248)
This commit includes a helm chart or Meli.  I created it based on the
k8s manifest contributed by @Berndinox.  Hopefully that `values.yaml`
file I provided sufficiently explains the various knobs to turn.  One
potential area of improvement would be allowing existing PVCs to
be used.

Note that this includes the feature I mentioned in #233 but which
also applies to #238 which is the ability to explicitly list all sites
so that the ingress configuration can fetch SSL certificates
_for branch host names_.

Ideally, this helm chart would be packaged up and shared...somewhere.
I'm actually not that well versed in how to host helm charts for
open source software.  I know there used to be a public registry
hosted by Google but then they withdrew support and the
landscape got very fragmented.  If there _is_ a place to host such
a chart, it would be good to publish this there so that an ordinary
helm installation could reference it.

Not that the `values.yaml` file include the image and tag.  This
can be customized to run a custom Meli image instead of the
official Docker hub image.

This closes #246.

Co-authored-by: Michael Tiller <michael.tiller@ricardo.com>