feat(ui): Improve toast error messages
This just modified the error handling on several axios calls
to extract the underlying error embedded in the HTTP response
so it is visible to the user.  The function for doing this already
existed but was never used (and had a typo in the name which
I've also corrected here).